requirements.txt
pytest
pytest_bdd==3.2.1
selenium
jsonpickle
pony
pymysql
mysql-connector-python-rf
allure-pytest
